// client/src/domain/sdcForm/SDCForm.ts
export type SDCQuestion = ResponseQuestion | SelectQuestion;

export type SDCQuestionProps = ParentQuestionProps & {
  question: string;
  questionId: string;

  isRequired: boolean;
  /**
   * The minimum number of repetitions allowed for a Question.
   *
   * If 0, then the question and all descendent questions are optional to answer.
   * @default 1
   */
  minRepetitions: number;
  /**
   * The maximum number of repetitions allowed for a Question.
   *
   * If 0, infinitely repeating
   * @default 1
   */
  maxRepetitions: number;
};
// ...
export type ParentQuestionProps =
  | {
      isParentQuestion: 'true';
      subQuestions: SDCQuestion[];
    }
  | {
      isParentQuestion: 'false';
    };
// ...
export type SelectQuestion = SDCQuestionProps & {
  type: 'SelectQuestion';
  /**
   * If 0, unlimited
   * if 1, single select
   * if > 1, max selections is finite but not 1.
   */
  maxSelections: number;
  /**
   * Minimum amount of selections for this select question.
   * Minimum value is 1. Only applicable when maxSelections is not 0.
   */
  minSelections: number;
  /**
   * There may be 0 options. In the case of 0 options, there may actually
   *  be `subQuestions` as part of the Question.
   *
   * If there are 0 options, render nothing for the options and render the subQuestions
   *  if they exist.
   *
   * An example of 0 can be found in: `PKG_THYROID_US.json`. Search for "Right Lobe"
   *  in the document and notice that there are 0 options but there are still subQuestions
   *  to render.
   */
  options: SDCSelectOption[];
};

export type SDCSelectOption = {
  optionId: string;
  text: string;
  isInitiallySelected: boolean;

  /**
   * These can be populated if the question is a control question.
   */
  controlledQuestions?: SDCQuestion[];
} & (SDCSelectChoiceProps | SDCSelectResponseProps);

export interface SDCSelectChoiceProps {
  responseMode: 'choice';
}

export interface SDCSelectResponseProps {
  responseMode: 'responseText';
  responseType: SDCResponseQuestionBody;
  /** Whether the text in this response choice is required. */
  isRequired: boolean;
  responseSuffix?: string;
}